backend/requirements_25-09-2025.txt
setuptools==75.8.0
affine==2.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==2.4.1
attrs==23.2.0
blessings==1.7
blinker==1.9.0
branca==0.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contextily==1.6.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
python-dotenv==1.1.1
earthengine-api==1.7.18
ee==0.2
eerepr==0.0.4
executing==2.0.1
fiona==1.9.6
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.1
folium==0.16.0
fonttools==4.51.0
future==1.0.0
geedim==1.7.2
geemap==0.32.1
geocoder==1.38.1
geographiclib==2.0
geopandas==1.1.3
geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.18.2
ipython==8.32.0
ipytree==0.2.2
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.7
langchain==0.3.27
langchain-core==0.3.76
langchain-ollama==0.3.8
langchain-text-splitters==0.3.11
langsmith==0.4.31
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mercantile==1.2.1
mistune==3.1.4
numpy>=2.3.3
ollama==0.6.0
orjson==3.11.3
packaging==24.2
pandas==3.0.1
parso==0.8.4
pillow==12.1.1
plotly==6.5.2
prompt-toolkit==3.0.50
proto-plus==1.26.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyCRS==1.0.2
pydantic==2.12.5
Pygments==2.19.1
pyparsing==3.2.1
pyperclip==1.9.0
pyproj==3.6.1
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rasterio==1.4.4
rasterstats==0.19.0
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scooby==0.10.0
shapely==2.0.7
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
snuggs==1.4.7
SQLAlchemy>=2.0.48
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
xyzservices==2025.1.0
zstandard==0.25.0
gunicorn==23.0.0
google-generativeai==0.8.3